Add evaluation after each training epoch

+++ deeplima/libs/tasks/graph_dp/train/conllu_file_iterator.h
@@ -65,14 +65,14 @@ public:
     }
 
   public:
-    virtual void set_batch_size(size_t batch_size);
+    virtual void set_batch_size(int64_t batch_size);
     virtual void start_epoch();
     virtual bool end();
     virtual const Batch next_batch();
 
   private:
     const CoNLLUDataSet& m_dataset;
-    size_t m_batch_size;
+    int64_t m_batch_size;
 
     size_t m_current_bucket;
     size_t m_batch_start_offset;
